deployment/runpod/requirements.txt
runpod==1.6.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ultralytics>=8.0.0
onnxruntime-gpu>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface-hub>=0.20.0
accelerate>=0.25.0
safetensors>=0.4.0